abstract away hook
